kinopoisk_autotests/requirements.txt
pytest==7.4.0
selenium==4.8.0
requests==2.28.1
allure-pytest==2.13.2
pytest-xdist==3.2.0
webdriver-manager==3.8.5
pylint==2.17.4
flake8==6.0.0